Adicionado sistema de colisões.
1. Adicionado sistema de colisão entre entidades.
2. Adicionado movimento inicial randômico de Ball.
3. Alterado configurações das entidades.
4. Alterado tipo de algumas variáveis.
5. Adicionado retorno de valores das variáveis privadas de Game.

+++ src/Entity.java
@@ -4,17 +4,17 @@ import java.awt.Rectangle;
 
 public class Entity
 {
-	protected int posX;
-	protected int posY;
+	protected double posX;
+	protected double posY;
 	protected int width;
 	protected int height;
-	protected int speed;	
+	protected double speed;	
 	protected Rectangle bounds;
 	
 	public void render(Graphics graphic)
 	{
 		graphic.setColor(Color.WHITE);
-		graphic.fillRect(posX, posY, width, height);
-		bounds = new Rectangle(posX, posY, width, height);
+		graphic.fillRect((int)posX, (int)posY, width, height);
+		bounds = new Rectangle((int)posX, (int)posY, width, height);
 	}
 }
